Change BlockTimestamp internal integer to u64

+++ chainstate/src/detail/pow/helpers.rs
@@ -96,8 +96,8 @@ pub mod special_rules {
     /// Checks if it took > 20 minutes to find a block
     pub fn block_production_stalled(
         target_spacing_in_secs: u64,
-        new_block_time: u32,
-        prev_block_time: u32,
+        new_block_time: u64,
+        prev_block_time: u64,
     ) -> bool {
         new_block_time as u64 > (prev_block_time as u64 + (target_spacing_in_secs * 2))
     }
